[FE] Refactory: mutable한값을 immutable하게 변경 & 파일명 수정
